services/protocol-gateway/requirements.txt
# =============================================================================
# services/protocol-gateway/requirements.txt
# =============================================================================
#
# WHAT IS THIS FILE?
# This file lists all the Python packages that the Protocol Gateway needs
# to run. When pip reads this file, it downloads and installs each package
# and all of its dependencies automatically.
#
# HOW TO INSTALL:
#   pip install -r requirements.txt
#
# This is called by the Dockerfile during docker image build, and also
# by the LXC setup script when deploying on Proxmox.
#
# WHY PIN VERSIONS (e.g. pika==1.3.2)?
# Pinning to exact versions (==) means everyone running DAWS_BAS gets
# the exact same package versions. This prevents "it worked last month
# but broke today" situations caused by packages releasing new versions
# with breaking changes.
#
# The trade-off is that you don't automatically get bug fixes and security
# patches — you need to manually update version numbers when needed.
# For a learning project this is fine. For production, use tools like
# Dependabot to automate version updates.
#
# =============================================================================


# pika - RabbitMQ client library (AMQP protocol)
# Used in: publisher.py
# Handles connecting to RabbitMQ, declaring exchanges, and publishing messages
# AMQP (Advanced Message Queuing Protocol) is the protocol RabbitMQ speaks
pika==1.3.2


# BAC0 - BACnet client/server library
# Used in: bacnet.py
# High-level BACnet library built on top of bacpypes3
# Handles WhoIs/IAm discovery, device connections, and point reading
# We use this same library in the Traffic Light simulator
BAC0==2025.9.15


# bacpypes3 - Low-level BACnet protocol stack
# Used by: BAC0 internally
# BAC0 is built on top of bacpypes3 which handles the raw BACnet packet
# encoding/decoding, UDP networking, and async event loop integration
# We don't use bacpypes3 directly but it must be installed for BAC0 to work
bacpypes3==0.0.104


# pymodbus - Modbus TCP/RTU client library
# Used in: modbus.py
# Handles Modbus TCP connections, register reads (holding, input, coil)
# Supports both sync and async operation — we use the async version
pymodbus==3.7.4


# paho-mqtt - MQTT client library
# Used in: mqtt_client.py
# The official Eclipse Paho MQTT client for Python
# Handles connecting to MQTT brokers, subscribing to topics,
# and receiving published messages via callbacks
paho-mqtt==2.1.0


# asyncua - OPC-UA client library (async version)
# Used in: opcua.py
# Handles connecting to OPC-UA servers, browsing node trees,
# creating subscriptions, and receiving value change notifications
# "asyncua" is the async version of the "opcua" library (uses async/await)
asyncua==1.1.5


# aiohttp - Async HTTP client library
# Used in: lon.py
# Used to make HTTP GET requests to the LON SmartServer REST API
# "aio" prefix means it works with Python's asyncio (async/await)
# Similar to the popular "requests" library but async-compatible
aiohttp==3.9.5


# python-dotenv - Environment variable loader
# Used in: main.py (indirectly)
# Reads key=value pairs from a .env file and loads them as environment variables
# This lets you run the gateway locally without setting env vars manually —
# just create a .env file with your settings
python-dotenv==1.0.0
